feat: Implentation of missing DAO classes and methods

+++ src/main/java/com/model/Fines/InfractionType.java
@@ -1,14 +1,14 @@
 package com.model.Fines;
 
-import com.model.Automobile.Automobile;
+import java.math.BigDecimal;
 
 public class InfractionType {
     private String name;
     private String description;
-    private double amount;
+    private BigDecimal amount;
     private int scoring;
 
-    public InfractionType(String name,String description, double amount, int scoring) {
+    public InfractionType(String name,String description, BigDecimal amount, int scoring) {
         this.name = name;
         this.description = description;
         this.amount = amount;
@@ -18,12 +18,12 @@ public class InfractionType {
     //getters
     public String getName() { return name; }
     public String getDescription() { return description; }
-    public double getAmount() { return amount; }
+    public BigDecimal getAmount() { return amount; }
     public int getScoring() { return scoring; }
 
     //setters
     public void setDescription(String description) { this.description = description; }
-    public void setAmount(double amount) { this.amount = amount; }
+    public void setAmount(BigDecimal amount) { this.amount = amount; }
     public void setScoring(int scoring) { this.scoring = scoring; }
 
 }
